\name{all.datasets}
\docType{data}
\alias{all.datasets}
\title{data objects for rat liver_brain samples}
\description{
 A data frame providing the expression profilings of GSE19830 microarray samples.   
}
\usage{all.datasets}
\format{
 A matrix with expression studies in the GSE19830 microarray samples: the first three columns are corresponding to the liver, while the last three samples are corresponding to the brain.    
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(rat_liver_brain)
}


