\name{array.proportions}
\docType{data}
\alias{array.proportions}
\title{proportions for rat liver and brain mixing samples}
\description{
 array.proportions: a data frame providing the fractions for liver and brain from the microarray GSE 19830 study
}
\usage{array.proportions}
\format{
 a martix whose rows are mixing samples' name and columns are fractions from pure live and brain tissues
}
\author{
       Ting Gong \email{tinggong@gmail.com}
       Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(rat_liver_brain)
}