\name{array.signatures}
\docType{data}
\alias{array.signatures}
\title{data objects for rat liver and brain pure samples}
\description{
 array.signatures: a data frame providing the expression values from rat pure liver and brain samples, each has threee replicates
}
\usage{array.signatures}
\format{
 a data matrix with 30 expressions from rat pure liver and brain tissues
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(rat_liver_brain)
}


