\name{datasets}
\docType{data}
\alias{datasets}
\title{data objects for liver and kidney mixing samples}
\description{
 A data drame providing the RPKM of seven mixing samples.       
}
\usage{datasets}
\format{
 A data frame with 31979 genes' expression on the 7 mixing samples: reads.1, reads.2, reads.3, reads.4, reads.5, reads.6, reads.7
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(liver_kidney)
}


