\name{fraction}
\docType{data}
\alias{fraction}
\title{mixing fractions for multi-tissues mixing samples}
\description{
 A data frame providing the fractions from 5 tissues in the mixing samples
}
\usage{fraction}
\format{
 A martix whose rows are mixing samples' name and columns are fractions from pure tissues including brain, muscle, lung, liver and heart
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(multi_tissue)
}


