\name{proportions}
\docType{data}
\alias{proportions}
\title{proportions for liver and kidney mixing samples}
\description{
 proportions: a data frame providing the fractions for liver and kidney in the mixing samples
}
\usage{proportions}
\format{
 a martix whose rows are mixing samples' name and columns are fractions from pure live and kidney tissues
}
\author{
       Ting Gong \email{tinggong@gmail.com}
       Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(liver_kidney)
}


