\name{signatures}
\docType{data}
\alias{signatures}
\title{data objects for liver and kidney pure samples}
\description{
 signatures: a data frame providing the expression values from pure liver and kidney samples
}
\usage{signatures}
\format{
 a data matrix with 630 expressions from pure liver and kidney tissues
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(liver_kidney)
}


