\name{x.data}
\docType{data}
\alias{x.data}
\title{data objects for multi-tissues mixing samples}
\description{
 A data frame providing the RPKM of nine mixing samples.   
}
\usage{x.data}
\format{
 A matrix with all the genes' expression in the mixing samples: the first two columns are corresponding to the RefSeq accession numbers and gene symbols   
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(multi_tissue)
}


