\name{x.signature}
\docType{data}
\alias{x.signature}
\title{data objects for multi-tissues pure samples}
\description{
 A data frame providing the expression values from pure brain, muscle, lung, liver and heart samples.
}
\usage{x.signature}
\format{
 A martix whose rows are gene symbols and columns are RPKM expressions from pure tissues.
}
\author{Ting Gong \email{tinggong@gmail.com}
        Joseph D. Szustakowski \email{joseph.szustakowski@novartis.com}
}
\keyword{datasets}
\examples{
data(multi_tissue)
}


